Add back start frame to show()
The start frame had previously been removed, because we're already
sending a start frame as part of the end frame, which we do to stay
compatible with both versions of the APA102, combined with a start
frame during initialization. However, this doesn't take into account
removing and reconnecting an APA102 strip/board while the code is
already running, potentially missing a start frame.